requirements.txt
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==2.0.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
Thingy==0.8.5
urllib3==1.25.9
Werkzeug==0.16.1
zipp==3.1.0
